Fix ClassCategories converter
