+++ src/Eccube/Twig/Extension/EccubeExtension.php
@@ -24,6 +24,7 @@
 
 namespace Eccube\Twig\Extension;
 
+use Eccube\Entity\Product;
 use Eccube\Service\TaxRuleService;
 use Eccube\Util\StringUtil;
 use Symfony\Component\Form\FormView;
@@ -55,6 +56,7 @@ class EccubeExtension extends AbstractExtension
             new TwigFunction('is_object', array($this, 'isObject')),
             new TwigFunction('calc_inc_tax', array($this, 'getCalcIncTax')),
             new TwigFunction('active_menus', array($this, 'getActiveMenus')),
+            new TwigFunction('class_categories_as_json', array($this, 'getClassCategoriesAsJson')),
             new TwigFunction('php_*', function () {
                     $arg_list = func_get_args();
                     $function = array_shift($arg_list);
@@ -255,5 +257,53 @@ class EccubeExtension extends AbstractExtension
         trigger_error('Called to an undefined function : php_'.$function, E_USER_WARNING);
 
         return null;
+    }
+
+    /**
+     * Get the ClassCategories as JSON.
+     *
+     * @param Product $Product
+     * @return string
+     */
+    public function getClassCategoriesAsJson(Product $Product)
+    {
+        $Product->_calc();
+        $class_categories = [
+            '__unselected' => [
+                '__unselected' => [
+                    'name'              => trans('product.text.please_select'),
+                    'product_class_id'  => '',
+                ],
+            ],
+        ];
+        foreach ($Product->getProductClasses() as $ProductClass) {
+            /* @var $ProductClass \Eccube\Entity\ProductClass */
+            $ClassCategory1 = $ProductClass->getClassCategory1();
+            $ClassCategory2 = $ProductClass->getClassCategory2();
+            if ($ClassCategory2 && !$ClassCategory2->isVisible()) {
+                continue;
+            }
+            $class_category_id1 = $ClassCategory1 ? (string) $ClassCategory1->getId() : '__unselected2';
+            $class_category_id2 = $ClassCategory2 ? (string) $ClassCategory2->getId() : '';
+            $class_category_name2 = $ClassCategory2 ? $ClassCategory2->getName().($ProductClass->getStockFind() ? '' : trans('product.text.out_of_stock')) : '';
+
+            $class_categories[$class_category_id1]['#'] = array(
+                'classcategory_id2' => '',
+                'name'              => trans('product.text.please_select'),
+                'product_class_id'  => '',
+            );
+            $class_categories[$class_category_id1]['#'.$class_category_id2] = array(
+                'classcategory_id2' => $class_category_id2,
+                'name'              => $class_category_name2,
+                'stock_find'        => $ProductClass->getStockFind(),
+                'price01'           => $ProductClass->getPrice01() === null ? '' : number_format($ProductClass->getPrice01IncTax()),
+                'price02'           => number_format($ProductClass->getPrice02IncTax()),
+                'product_class_id'  => (string) $ProductClass->getId(),
+                'product_code'      => $ProductClass->getCode() === null ? '' : $ProductClass->getCode(),
+                'sale_type'      => (string) $ProductClass->getSaleType()->getId(),
+            );
+        }
+
+        return json_encode($class_categories);
     }
 }
