+++ tests/Eccube/Tests/Web/ProductControllerTest.php
@@ -221,57 +221,4 @@ class ProductControllerTest extends AbstractWebTestCase
         $this->assertContains('カートに入れる', $html);
         $this->assertContains('お気に入りに追加済みです', $html);
     }
-
-    public function testCreateClassCategories()
-    {
-        $Controller = new ProductController(
-            $this->container->get('eccube.purchase.flow.cart'),
-            $this->container->get(CustomerFavoriteProductRepository::class),
-            $this->container->get(CartService::class),
-            $this->container->get(ProductRepository::class),
-            $this->container->get(BaseInfo::class),
-            $this->container->get('security.authentication_utils')
-        );
-        $faker = $this->getFaker();
-        $Product = $this->createProduct($faker->word, 3);
-
-        $RefClass = new \ReflectionClass($Controller);
-        $RefMethod = $RefClass->getMethod('createClassCategoriesAsArray');
-        $RefMethod->setAccessible(true);
-        $actuals = $RefMethod->invoke($Controller, $Product);
-
-        foreach ($Product->getClassCategories1() as $class_category_id => $name) {
-            $this->assertArrayHasKey($class_category_id, $actuals);
-
-            $ClassCategory2 = $Product->getClassCategories2($class_category_id);
-            if (empty($ClassCategory2)) {
-                $this->markTestSkipped('ClassCategory2 is empty.');
-            }
-
-            foreach ($ClassCategory2 as $class_category_id2 => $name2) {
-                $this->assertArrayHasKey('#'.$class_category_id2, $actuals[$class_category_id]);
-
-                $actual = $actuals[$class_category_id]['#'.$class_category_id2];
-
-                $this->assertEquals($class_category_id2, $actual['classcategory_id2']);
-                $this->assertEquals($name2, $actual['name']);
-
-                $ProductClass = $Product
-                    ->getProductClasses()
-                    ->filter(
-                        function ($ProductClass) use ($actual) {
-                            return $ProductClass->getId() == $actual['product_class_id'];
-                        })
-                    ->first();
-
-                if ($ProductClass->getPrice01IncTax()) {
-                    $this->assertEquals(number_format($ProductClass->getPrice01IncTax()), $actual['price01']);
-                }
-                $this->assertEquals(number_format($ProductClass->getPrice02IncTax()), $actual['price02']);
-                $this->assertEquals($ProductClass->getCode(), $actual['product_code']);
-                $this->assertEquals($ProductClass->getSaleType()->getId(), $actual['sale_type']);
-                $this->assertEquals($ProductClass->getStockFind(), $actual['stock_find']);
-            }
-        }
-    }
 }
